requirements.txt
matplotlib==3.4.3
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
plotly==5.1.0
requests==2.26.0
streamlit==0.88.0
streamlit-option-menu==0.0.3
